feat: restructure and easier addition of new ETL
