+++ Dockerfile
@@ -3,6 +3,6 @@ FROM python:3.8-slim-buster
 RUN pip3 install --upgrade pip==20.0.*
 RUN pip3 install requests==2.23.*
 
-COPY ./johns_hopkins_etl.py /johns_hopkins_etl.py
+COPY ./covid19-etl /covid19-etl
 
-ENTRYPOINT [ "python3", "/johns_hopkins_etl.py" ]
+CMD [ "python3", "/covid19-etl/main.py" ]
