+++ covid19-etl/etl/idph.py
@@ -1,46 +1,11 @@
-import csv
 import datetime
-import enum
-import json
-import os
 import re
-from collections import defaultdict
 from contextlib import closing
-from math import ceil
 
 import requests
 
-BASE_URL = "https://covid19.datacommons.io"
-PROGRAM_NAME = "open"
-PROJECT_CODE = "IDPH"
-
-# Note: if we end up having too much data, Sheepdog submissions may
-# time out. We'll have to use a smaller batch size and hope that's enough
-SUBMIT_BATCH_SIZE = 100
-
-
-# with open("IL_counties_central_coords_lat_long.tsv") as f:
-#     counties = f.readlines()
-#     counties = counties[1:]
-#     counties = map(lambda l: l.strip().split("\t"), counties)
-
-# county_dict = {}
-# for county, lat, lon in counties:
-#     print(county)
-#     county_dict[county] = {'lat': lat, 'lon': lon}
-
-def main():
-    token = os.environ.get("ACCESS_TOKEN")
-    if not token:
-        raise Exception(
-            "Need ACCESS_TOKEN environment variable (token for user with read and write access to {}-{})".format(
-                PROGRAM_NAME, PROJECT_CODE
-            )
-        )
-
-    etl = IllinoisDPHETL(token)
-    etl.files_to_submissions()
-    etl.submit_metadata()
+from etl import base
+from helper.metadata_helper import MetadataHelper
 
 
 def format_summary_location_submitter_id(country, state, county):
@@ -59,18 +24,38 @@ def format_summary_report_submitter_id(location_submitter_id, date):
     return "{}_{}".format(location_submitter_id.replace("summary_location_", "summary_report_"), date)
 
 
-class IllinoisDPHETL:
-    def __init__(self, access_token):
+class IDPH(base.BaseETL):
+    def __init__(self, base_url, access_token):
+        super().__init__(base_url, access_token)
+
+        self.program_name = "open"
+        self.project_code = "IDPH"
+        self.metadata_helper = MetadataHelper(base_url=self.base_url,
+                                              program_name=self.program_name,
+                                              project_code=self.project_code,
+                                              access_token=access_token)
+
+        self.county_dict = {}
+
         self.summary_locations = []
         self.summary_reports = []
-        self.metadata_helper = MetadataHelper(access_token=access_token)
+
+    def il_counties(self):
+        with open("IL_counties_central_coords_lat_long.tsv") as f:
+            counties = f.readlines()
+            counties = counties[1:]
+            counties = map(lambda l: l.strip().split("\t"), counties)
+
+        county_dict = {}
+        for county, lat, lon in counties:
+            self.county_dict[county] = {'lat': lat, 'lon': lon}
 
     def files_to_submissions(self):
         """
         Reads JSON file and convert the data to Sheepdog records
         """
 
-        latest_submitted_date = self.metadata_helper.get_latest_submitted_data()
+        latest_submitted_date = self.metadata_helper.get_latest_submitted_data_idph()
         today = datetime.date.today()
         if latest_submitted_date == today:
             print("Nothing to submit: today and latest submitted date are the same.")
@@ -83,11 +68,11 @@ class IllinoisDPHETL:
         # they changed the URL on April 1, 2020
         if today > datetime.date(2020, 3, 31):
             url = "http://www.dph.illinois.gov/sitefiles/COVIDTestResults.json"
-        else:    
+        else:
             url = f"https://www.dph.illinois.gov/sites/default/files/COVID19/COVID19CountyResults{today_str}.json"
-        self.parse_file(state, url)
+        self.parse_file(latest_submitted_date, state, url)
 
-    def parse_file(self, state, url):
+    def parse_file(self, latest_submitted_date, state, url):
         """
         Converts a JSON files to data we can submit via Sheepdog. Stores the
         records to submit in `self.summary_locations` and `self.summary_reports`.
@@ -102,6 +87,10 @@ class IllinoisDPHETL:
         with closing(requests.get(url, stream=True)) as r:
             data = r.json()
             date = self.get_date(data)
+
+            if date == latest_submitted_date.strftime("%Y-%m-%d"):
+                print("Nothing to submit: today and latest submitted date are the same.")
+                return
 
             for county in data["characteristics_by_county"]["values"]:
                 summary_location, summary_report = self.parse_county(
@@ -128,18 +117,18 @@ class IllinoisDPHETL:
             "country_region": country,
             "county": county,
             "submitter_id": summary_location_submitter_id,
-            "projects": [{"code": PROJECT_CODE}],
+            "projects": [{"code": self.project_code}],
             "province_state": state,
         }
 
-        # if county in county_dict:
-        #     summary_location["latitude"] = county_dict[county]["lat"]
-        #     summary_location["longitude"] = county_dict[county]["lon"]
-        # else:
-        #     if county_json["lat"] != 0:
-        #         summary_location["latitude"] = str(county_json["lat"])
-        #     if county_json["lon"] != 0:
-        #         summary_location["longitude"] = str(county_json["lon"])
+        if county in self.county_dict:
+            summary_location["latitude"] = self.county_dict[county]["lat"]
+            summary_location["longitude"] = self.county_dict[county]["lon"]
+        else:
+            if county_json["lat"] != 0:
+                summary_location["latitude"] = str(county_json["lat"])
+            if county_json["lon"] != 0:
+                summary_location["longitude"] = str(county_json["lon"])
 
         summary_report_submitter_id = format_summary_report_submitter_id(
             summary_location_submitter_id, date
@@ -186,89 +175,3 @@ class IllinoisDPHETL:
             rep_record.update(rep)
             self.metadata_helper.add_record_to_submit(rep_record)
         self.metadata_helper.batch_submit_records()
-
-
-class MetadataHelper:
-    def __init__(self, access_token):
-        self.base_url = BASE_URL
-        self.headers = {"Authorization": "bearer " + access_token}
-        self.project_id = "{}-{}".format(PROGRAM_NAME, PROJECT_CODE)
-        self.records_to_submit = []
-
-    def get_latest_submitted_data(self):
-        """
-        Queries Peregrine for the existing `summary_report` data.
-
-        { summary_report (first: 1, project_id: <...>) { date } }
-
-        Returns the latest submitted date as Python "datetime.date"
-        """
-        print("Getting latest date from Peregrine...")
-        query_string = (
-            '{ summary_report (first: 1, order_by_desc: "date", project_id: "'
-            + self.project_id
-            + '") { submitter_id date } }'
-        )
-        response = requests.post(
-            "{}/api/v0/submission/graphql".format(self.base_url),
-            json={"query": query_string, "variables": None},
-            headers=self.headers,
-        )
-        assert (
-            response.status_code == 200
-        ), "Unable to query Peregrine for existing data: {}\n{}".format(
-            response.status_code, response.text
-        )
-        try:
-            query_res = json.loads(response.text)
-        except:
-            print("Peregrine did not return JSON")
-            raise
-
-        report = query_res["data"]["summary_report"][0]
-        latest_submitted_date = datetime.datetime.strptime(
-            report["date"], "%Y-%m-%d")
-        return latest_submitted_date.date()
-
-    def add_record_to_submit(self, record):
-        self.records_to_submit.append(record)
-
-    def batch_submit_records(self):
-        """
-        Submits Sheepdog records in batch
-        """
-        if not self.records_to_submit:
-            print("  Nothing new to submit")
-            return
-
-        n_batches = ceil(len(self.records_to_submit) / SUBMIT_BATCH_SIZE)
-        for i in range(n_batches):
-            records = self.records_to_submit[
-                i * SUBMIT_BATCH_SIZE: (i + 1) * SUBMIT_BATCH_SIZE
-            ]
-            print(
-                "  Submitting {} records: {}".format(
-                    len(records), [r["submitter_id"] for r in records]
-                )
-            )
-            # self.records_to_submit = []  # TODO remove
-            # return  # TODO remove
-
-            response = requests.put(
-                "{}/api/v0/submission/{}/{}".format(
-                    self.base_url, PROGRAM_NAME, PROJECT_CODE
-                ),
-                headers=self.headers,
-                data=json.dumps(records),
-            )
-            assert (
-                response.status_code == 200
-            ), "Unable to submit to Sheepdog: {}\n{}".format(
-                response.status_code, response.text
-            )
-
-        self.records_to_submit = []
-
-
-if __name__ == "__main__":
-    main()
