--- idph_etl/IL_counties_central_coords_lat_long.tsv
@@ -1,103 +1,0 @@
-County	latitude	longitude
-Adams	40.01463743	-91.16403919
-Alexander	37.18823481	-89.31529760
-Bond	38.89685178	-89.43812561
-Boone	42.30935860	-88.83970261
-Brown	39.95675680	-90.73701414
-Bureau	41.42199135	-89.55082406
-Calhoun	39.20372238	-90.64681556
-Carroll	42.10103211	-89.90276400
-Cass	39.95815563	-90.22019895
-Champaign	40.11547418	-88.22393938
-Christian	39.54422275	-89.25193749
-Clark	39.35274677	-87.76859720
-Clay	38.73548275	-88.47378976
-Clinton	38.60541566	-89.48244967
-Coles	39.53450203	-88.22985840
-Cook	42.21054750	-87.44774127
-Crawford	39.01476881	-87.76311275
-Cumberland	39.28350144	-88.25662689
-DeKalb	41.94237518	-88.78605999
-De Witt	40.18672530	-88.96980362
-Douglas	39.75874433	-88.26343231
-DuPage	41.87080601	-88.08230427
-Edgar	39.69275602	-87.72361162
-Edwards	38.38444824	-88.06321825
-Effingham	39.06194899	-88.56844221
-Fayette	38.97065449	-89.05185509
-Ford	40.50867063	-88.23501682
-Franklin	37.99057503	-88.93523089
-Fulton	40.46201658	-90.21505945
-Gallatin	37.79085779	-88.22181102
-Greene	39.35952015	-90.44455210
-Grundy	41.30171752	-88.42756081
-Hamilton	38.07678032	-88.56375776
-Hancock	40.39484831	-91.14185800
-Hardin	37.50779833	-88.30451584
-Henderson	40.76411629	-90.93002616
-Henry	41.33521986	-90.11856630
-Iroquois	40.76593561	-87.84203026
-Jackson	37.79165441	-89.39202690
-Jasper	38.99299007	-88.15797642
-Jefferson	38.28153949	-88.92642117
-Jersey	39.09080200	-90.34172494
-Jo Daviess	42.39200007	-90.22999736
-Johnson	37.46585328	-88.89936774
-Kane	41.93303824	-88.44796117
-Kankakee	41.13482679	-87.90509088
-Kendall	41.59453845	-88.45931308
-Knox	40.91048546	-90.19498317
-Lake	42.31739786	-87.97247151
-La Salle	41.37146044	-88.91388219
-Lawrence	38.71082539	-87.72036552
-Lee	41.74625724	-89.27659552
-Livingston	40.90273312	-88.54803352
-Logan	40.14004498	-89.35103014
-Macon	39.85376530	-88.94413037
-Macoupin	39.28582859	-89.93337042
-Madison	38.83989048	-89.86625617
-Marion	38.63001421	-88.92769909
-Marshall	41.03522205	-89.37160797
-Mason	40.20695031	-89.95679045
-Massac	37.22509066	-88.67878215
-McDonough	40.43975811	-90.65702295
-McHenry	42.31123317	-88.44631849
-McLean	40.48936810	-88.83059120
-Menard	40.02525690	-89.80713463
-Mercer	41.21257605	-90.70468331
-Monroe	38.28257980	-90.20769162
-Montgomery	39.23574972	-89.53835123
-Morgan	39.73900877	-90.20503139
-Moultrie	39.65545572	-88.60486330
-Ogle	42.05384102	-89.34614999
-Peoria	40.79130797	-89.77153397
-Perry	38.06969681	-89.36088117
-Piatt	39.95685005	-88.61022339
-Pike	39.65449535	-90.85879517
-Pope	37.41243331	-88.54689650
-Pulaski	37.22004700	-89.15041542
-Putnam	41.20679964	-89.29606171
-Randolph	38.05865009	-89.80139966
-Richland	38.70514584	-88.10141373
-Rock Island	41.45407335	-90.41831970
-St. Clair	38.49255625	-89.90962855
-Saline	37.73336874	-88.55969838
-Sangamon	39.78635049	-89.68066635
-Schuyler	40.15919215	-90.62091719
-Scott	39.65412222	-90.46587753
-Shelby	39.37001801	-88.78817283
-Stark	41.09865025	-89.79304822
-Stephenson	42.36800835	-89.69983128
-Tazewell	40.49083478	-89.52366045
-Union	37.47506256	-89.22056770
-Vermilion	40.17876668	-87.72541046
-Wabash	38.43337904	-87.87070629
-Warren	40.81915392	-90.60302141
-Washington	38.35988315	-89.37735094
-Wayne	38.41579151	-88.46648788
-White	38.08231869	-88.17190743
-Whiteside	41.75719511	-89.91921725
-Will	41.45104497	-88.03289760
-Williamson	37.72472254	-88.94911385
-Winnebago	42.35512218	-89.18309307
-Woodford	40.79326894	-89.20831081
